Separate widget test
